requirements.txt
Flask==2.0.1
newspaper3k==0.2.8
elasticsearch==7.13.4
flask_paginate==0.9.0
